Split runner artifact upload tests
